# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NCH Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        notes:
          type: string
        attendees:
          type: array
          items:
            type: string
        linkedJobId:
          type: integer
        linkedJobNumber:
          type: string
        createdBy:
          type: string
        ownerRole:
          type: string
          enum: [mike, jack, jacob]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - startTime
        - ownerRole
        - createdBy
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        notes:
          type: string
        attendees:
          type: array
          items:
            type: string
        linkedJobId:
          type: integer
      required:
        - title
        - startTime
    UpdateAppointmentBody:
      type: object
      properties:
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        notes:
          type: string
        attendees:
          type: array
          items:
            type: string
        linkedJobId:
          type: integer
...